Setting APPLICATION_EXTENSION_API_ONLY to YES
Per the documentation:

```
When enabled, this causes the compiler and linker to
disallow use of APIs that are not available
to app extensions and to disallow linking to frameworks
that have not been built with this setting enabled.
```

Since we're not using any of these APIs, we can safely turn this setting on.