Enable free materialize function and use NSError
That bug isn't fixed yet, but this also won't ever compile. Somewhat
Reasonably, you are unable to use a protocol as a concrete type for a
generic type that has a type constraint with that protocol.

To get around this, we should use `NSError` for the error type, since we
can always safely cast from `ErrorType` to `NSError`.