Reconsider ErrorTypeConvertible
This was changed in https://github.com/antitypical/Result/pull/141. But the current form does not looks correct since the `ConvertibleType` associated type is not used in any method signature.

This change brings back the original definition and worked well for both Swift 2.2 on OS X and swift-DEVELOPMENT-SNAPSHOT-2016-03-24-a on Linux.